# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for JAV metadata scraping and file organization",
        "title": "Javinizer API",
        "termsOfService": "https://github.com/javinizer/javinizer-go",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/javinizer/javinizer-go/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/jobs/{id}/operations/{movieId}/revert": {
            "post": {
                "description": "Revert file operations for a specific movie within a batch job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Revert a specific movie within a batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.RevertResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Revert is disabled",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}/revert": {
            "post": {
                "description": "Revert all file organization operations for a batch job, moving files back to original paths",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Revert a batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.RevertResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Revert is disabled",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Movie not found"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.RevertFileError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "file not found"
                },
                "movie_id": {
                    "type": "string",
                    "example": "ABC-123"
                },
                "new_path": {
                    "type": "string",
                    "example": "/dest/ABC-123 [Studio]/ABC-123.mp4"
                },
                "operation_id": {
                    "type": "integer",
                    "example": 5
                },
                "original_path": {
                    "type": "string",
                    "example": "/source/ABC-123.mp4"
                },
                "outcome": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.RevertOutcomeEnum"
                        }
                    ],
                    "example": "skipped"
                },
                "reason": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.RevertReasonEnum"
                        }
                    ],
                    "example": "anchor_missing"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.RevertResultResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.RevertFileError"
                    }
                },
                "failed": {
                    "type": "integer",
                    "example": 1
                },
                "job_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "skipped": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus"
                        }
                    ],
                    "example": "reverted"
                },
                "succeeded": {
                    "type": "integer",
                    "example": 9
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_models.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "cancelled",
                "organized",
                "reverted"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed",
                "JobStatusCancelled",
                "JobStatusOrganized",
                "JobStatusReverted"
            ]
        },
...
        "github_com_javinizer_javinizer-go_internal_models.RevertOutcomeEnum": {
            "type": "string",
            "enum": [
                "reverted",
                "skipped",
                "failed"
            ],
            "x-enum-comments": {
                "RevertOutcomeFailed": "Failed to revert",
                "RevertOutcomeReverted": "Successfully reverted",
                "RevertOutcomeSkipped": "Skipped (e.g., anchor missing)"
            },
            "x-enum-descriptions": [
                "Successfully reverted",
                "Skipped (e.g., anchor missing)",
                "Failed to revert"
            ],
            "x-enum-varnames": [
                "RevertOutcomeReverted",
                "RevertOutcomeSkipped",
                "RevertOutcomeFailed"
            ]
        },
        "github_com_javinizer_javinizer-go_internal_models.RevertReasonEnum": {
            "type": "string",
            "enum": [
                "anchor_missing",
                "destination_conflict",
                "access_denied",
                "unexpected_path_state",
                "nfo_restore_failed",
                "generated_cleanup_failed"
            ],
            "x-enum-comments": {
                "RevertReasonAccessDenied": "Permission error during revert",
                "RevertReasonAnchorMissing": "Video file missing at expected path",
                "RevertReasonDestinationConflict": "Original path already occupied",
                "RevertReasonGeneratedCleanupFailed": "Generated file cleanup failed",
                "RevertReasonNFORestoreFailed": "NFO write failed",
                "RevertReasonUnexpectedPathState": "File in unexpected state"
            },
            "x-enum-descriptions": [
                "Video file missing at expected path",
                "Original path already occupied",
                "Permission error during revert",
                "File in unexpected state",
                "NFO write failed",
                "Generated file cleanup failed"
            ],
            "x-enum-varnames": [
                "RevertReasonAnchorMissing",
                "RevertReasonDestinationConflict",
                "RevertReasonAccessDenied",
                "RevertReasonUnexpectedPathState",
                "RevertReasonNFORestoreFailed",
                "RevertReasonGeneratedCleanupFailed"
            ]
        },
...
    }
}